fix icon  and change the name file icon
