+++ apps/core/app/dashboard/page.tsx
@@ -1,8 +1,8 @@
 import { Input } from "@repo/ui/components/input";
 import Serchinicon from "../../../../packages/icons/src/components/serach";
 import Bellicon from "../../../../packages/icons/src/components/bell";
-import Chevrondownicon from "../../../../packages/icons/src/components/chevrondown";
-import Infosquareroundedicon from "../../../../packages/icons/src/components/infosquarerounded";
+import Chevrondownicon from "../../../../packages/icons/src/components/chevron-down";
+import Infosquareroundedicon from "../../../../packages/icons/src/components/info-square-rounded";
 import { Button } from "@repo/ui/components/button";
 import { Label } from "@repo/ui/components/label";
 import { Textarea } from "@repo/ui/components/textarea";
@@ -14,7 +14,7 @@ import {
   TabsTrigger,
 } from "@repo/ui/components/tabs";
 import { Switch } from "@repo/ui/components/switch";
-import Infocircleicon from "@repo/icons/infocircle";
+import Infocircleicon from "@repo/icons/info-circle";
 
 const DashboardPage = () => {
   return (
@@ -417,6 +417,7 @@ const DashboardPage = () => {
               <Button variant="secondary" className="px-4 py-2 bg-primary-600">
                 Submit for review
               </Button>
+              
             </div>
           </div>
         </div>
