+++ packages/icons/package.json
@@ -16,17 +16,17 @@
     "./figma": "./src/components/figma.tsx",
     "./adobexd": "./src/components/adobexd.tsx",
     "./sketch": "./src/components/sketch.tsx",
-    "./chevronright": "./src/components/chevronright.tsx",
-    "./messagecircle": "./src/components/messagecircle.tsx",
+    "./chevron-right": "./src/components/chevron-right.tsx",
+    "./message-circle": "./src/components/message-circle.tsx",
     "./heart1": "./src/components/heart1.tsx",
-    "./shoppingcartplus": "./src/components/shoppingcartplus.tsx",
-    "./Circlecheck": "./src/components/circlecheck.tsx",
-    "./clouddownload": "./src/components/clouddownload.tsx",
+    "./shopping-cart-plus": "./src/components/shopping-cart-plus.tsx",
+    "./Circle-check": "./src/components/circle-check.tsx",
+    "./cloud-download": "./src/components/cloud-download.tsx",
     "./search": "./src/components/search.tsx",
     "./bell": "./src/components/bell.tsx",
-    "./chevrondown": "./src/components/chevrondown.tsx",
-    "./infosquarerounded": "./src/components/infosquarerounded.tsx",
-    "./infocircle" : "./src/components/infocircle.tsx",
+    "./chevron-down": "./src/components/chevron-down.tsx",
+    "./info-square-rounded": "./src/components/info-square-rounded.tsx",
+    "./info-circle": "./src/components/info-circle.tsx",
     "./types": "./src/types/types.ts"
   },
   "scripts": {
