+++ packages/icons/src/components/adobexd.tsx
@@ -5,8 +5,8 @@ const Adobexdicon = (props: IconProps) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
+      width={size}
+      height={size}
       viewBox="0 0 16 16"
       fill="none"
       {...resProps}
