+++ packages/icons/src/components/chevron-down.tsx
@@ -1,12 +1,12 @@
 import { IconProps } from "../types/types";
-const Circlecheckicon = (props: IconProps) => {
+const Chevrondownicon = (props: IconProps) => {
   const { size = 24, color = "currentColor", ...resProps } = props;
 
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
@@ -16,10 +16,9 @@ const Circlecheckicon = (props: IconProps) => {
       {...resProps}
     >
       <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-      <path d="M12 12m-9 0a9 9 0 1 0 18 0a9 9 0 1 0 -18 0" />
-      <path d="M9 12l2 2l4 -4" />
+      <path d="M6 9l6 6l6 -6" />
     </svg>
   );
 };
 
-export default Circlecheckicon;
+export default Chevrondownicon;
