+++ packages/icons/src/components/chevron-right.tsx
@@ -5,8 +5,8 @@ const Chevronrighticon = (props: IconProps) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 24 24"
       fill="none"
       {...resProps}
