+++ packages/icons/src/components/circle-check.tsx
@@ -1,12 +1,12 @@
 import { IconProps } from "../types/types";
-const Infosquareroundedicon = (props: IconProps) => {
+const Circlecheckicon = (props: IconProps) => {
   const { size = 24, color = "currentColor", ...resProps } = props;
 
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
@@ -16,11 +16,10 @@ const Infosquareroundedicon = (props: IconProps) => {
       {...resProps}
     >
       <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-      <path d="M12 9h.01" />
-      <path d="M11 12h1v4h1" />
-      <path d="M12 3c7.2 0 9 1.8 9 9s-1.8 9 -9 9s-9 -1.8 -9 -9s1.8 -9 9 -9z" />
+      <path d="M12 12m-9 0a9 9 0 1 0 18 0a9 9 0 1 0 -18 0" />
+      <path d="M9 12l2 2l4 -4" />
     </svg>
   );
 };
 
-export default Infosquareroundedicon;
+export default Circlecheckicon;
