+++ packages/icons/src/components/info-circle.tsx
@@ -1,12 +1,12 @@
 import { IconProps } from "../types/types";
-const Chevrondownicon = (props: IconProps) => {
+const Infocircleicon = (props: IconProps) => {
   const { size = 24, color = "currentColor", ...resProps } = props;
 
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
@@ -16,9 +16,11 @@ const Chevrondownicon = (props: IconProps) => {
       {...resProps}
     >
       <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-      <path d="M6 9l6 6l6 -6" />
+      <path d="M3 12a9 9 0 1 0 18 0a9 9 0 0 0 -18 0" />
+      <path d="M12 9h.01" />
+      <path d="M11 12h1v4h1" />
     </svg>
   );
 };
 
-export default Chevrondownicon;
+export default Infocircleicon;
