+++ packages/icons/src/components/linkedin.tsx
@@ -4,8 +4,8 @@ const Linkedinicon = (props: IconProps) => {
 
   return (
     <svg
-      width="24"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 24 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
