+++ packages/icons/src/components/message-circle.tsx
@@ -5,8 +5,8 @@ const Messagecircleicon = (props: IconProps) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="28"
-      height="28"
+      width={size}
+      height={size}
       viewBox="0 0 28 28"
       fill="none"
       {...resProps}
