+++ packages/icons/src/components/shopping-cart-plus.tsx
@@ -5,8 +5,8 @@ const Shopingcartplusicon = (props: IconProps) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
-      width="25"
-      height="24"
+      width={size}
+      height={size}
       viewBox="0 0 25 24"
       fill="none"
       {...resProps}
