+++ packages/icons/src/components/youtube.tsx
@@ -4,8 +4,8 @@ const YouTubeIcon = (props: IconProps) => {
 
   return (
     <svg
-      width="19"
-      height="14"
+      width={size}
+      height={size}
       viewBox="0 0 19 14"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
